+++ packages/expo-location/CHANGELOG.md
@@ -15,6 +15,7 @@
 ### 🎉 New features
 
 - [Android] Make foreground service permission opt-in with `isAndroidForegroundServiceEnabled` config plugin option [#27265](https://github.com/expo/expo/pull/27265) by [@brentvatne](https://github.com/brentvatne))
+- [Android] Enable foreground service by default when background location is enabled [#27359](https://github.com/expo/expo/pull/27359) by [@brentvatne](https://github.com/brentvatne))
 
 ## 16.5.3 — 2024-02-06